// web-ui/src/types/commands.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LoopMode } from "./loop_mode";
import type { TrackId } from "./track_id";

export type Command = { "Queue": QueueCommand } | { "Playback": PlaybackCommand } | { "Library": LibraryCommand };

export type LibraryCommand = "Fetch";

export type PlaybackCommand = "Play" | "Pause" | "TogglePlay" | { "SetPosition": { position: { secs: number, nanos: number }, } } | { "Seek": { offset_seconds: bigint, } } | "Restart" | { "IncreaseVol": { step: number, } } | { "DecreaseVol": { step: number, } } | { "SetVolume": { value: number, } } | "GetSnapshot";

export type QueueCommand = { "Enqueue": { track_id: TrackId, } } | { "Prepend": { track_id: TrackId, } } | { "Dequeue": { index: number, } } | { "PlayNow": { track_id: TrackId, } } | "Next" | "Prev" | "Shuffle" | "Clear" | { "SetLoopMode": LoopMode } | "CycleLoopMode" | "Fetch";