// web-ui/src/types/server_event.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlayerSnapshot } from "./player_snapshot";
import type { QueueSnapshot } from "./queue_snapshot";
import type { Track } from "./track";
import type { TrackId } from "./track_id";

export type ServerEvent = { "type": "SendLibrary", "data": { [key in TrackId]: Track } } | { "type": "SendQueue", "data": QueueSnapshot } | { "type": "SendPlayerSnapshot", "data": PlayerSnapshot } | { "type": "SendPlayerPosition", "data": { position: { secs: number, nanos: number }, } };